libmesh_transform_reduce -> parallel_algorithms.h
We don't want to pull <numeric> into every single libMesh file... yet!
